+++ src/covert_wrappers.h
@@ -68,7 +68,7 @@ struct upseudo_header
     u_int16_t udp_length;
 };
 
-char covert_udp_recv(char *sip, unsigned short sport, bool ttl, bool tos, bool ipid) ;
+char covert_udp_recv(char *sip, int sport, bool ttl, bool tos, bool ipid) ;
 void covert_udp_send_data(char *sip, char *dip, unsigned short sport, unsigned short dport, char* data, int covert_channel);
 void covert_udp_send(char *sip, char *dip, unsigned short sport, unsigned short dport, unsigned char* data, int covert_channel);
 void recv_results(char* sip, unsigned short sport, char* filename,bool tcp);
